read all setting from a setting.txt that wise will write to. a default in buildartifacts that will be used in develop environment

+++ Teleopti.Support.Code/Tool/RefreshConfigFile.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 
 namespace Teleopti.Support.Code.Tool
 {
@@ -11,23 +12,24 @@ namespace Teleopti.Support.Code.Tool
             _configFileTagReplacer = configFileTagReplacer;
         }
 
-        public void ReplaceFile(string oldFilePath, string newFilePath)
+        public void ReplaceFile(string oldFilePath, string newFilePath, IList<SearchReplace> searchReplaces, bool doNotOverWrite)
         {
+            if (System.IO.File.Exists(oldFilePath) && doNotOverWrite ) return;
             System.IO.File.Copy(newFilePath, oldFilePath,true);
-            _configFileTagReplacer.ReplaceTags(oldFilePath);
+            _configFileTagReplacer.ReplaceTags(oldFilePath,searchReplaces);
         }
 
-        public void SplitAndReplace(string oldAndNewFile)
+        public void SplitAndReplace(string oldAndNewFile, IList<SearchReplace> searchReplaces, bool doNotOverWrite)
         {
             var files = oldAndNewFile.Split(',');
             if (files.GetUpperBound(0).Equals(1))
-                ReplaceFile(files[0], files[1]);
+                ReplaceFile(files[0], files[1], searchReplaces, doNotOverWrite);
         }
 
-        public void ReadLinesFromString(string allFilesToRefresh)
+        public void ReadLinesFromString(string allFilesToRefresh,IList<SearchReplace> searchReplaces, bool doNotOverWrite)
         {
             foreach (var myString in allFilesToRefresh.Split(new[] { Environment.NewLine }, StringSplitOptions.RemoveEmptyEntries))
-                SplitAndReplace(myString);
+                SplitAndReplace(myString, searchReplaces, doNotOverWrite);
         }
     }
 }
